pyinstaller missing binary fixed

+++ main.py
@@ -4,34 +4,42 @@ import cv2
 
 def video2thumb_by_time(file_path, time, output_file):
     try:
-        # read video
-        vcap = cv2.VideoCapture(file_path)
+        if os.path.exists(file_path):
+            # read video
+            vcap = cv2.VideoCapture(file_path)
 
-        fps_int = round(vcap.get(cv2.CAP_PROP_FPS))  # get fps
-        total_length = vcap.get(cv2.CAP_PROP_FRAME_COUNT) / \
-            vcap.get(cv2.CAP_PROP_FPS)
+            fps_int = round(vcap.get(cv2.CAP_PROP_FPS))  # get fps
+            try:
+                total_length = round(
+                    vcap.get(cv2.CAP_PROP_FRAME_COUNT) / vcap.get(cv2.CAP_PROP_FPS))
+            except:
+                total_length = round(vcap.get(cv2.CAP_PROP_FRAME_COUNT) / 30)
+            logging.info('length: {} -- fps: {}'.format(total_length, fps_int))
+            # print('video solution {}x{}'.format(
+            #     vcap.get(cv2.CAP_PROP_FRAME_WIDTH), vcap.get(cv2.CAP_PROP_FRAME_HEIGHT)))
 
-        # print('video solution {}x{}'.format(
-        #     vcap.get(cv2.CAP_PROP_FRAME_WIDTH), vcap.get(cv2.CAP_PROP_FRAME_HEIGHT)))
+            # if time surpass total length then time = 1 (s)
+            if time > total_length:
+                time = 1
 
-        # if time surpass total length then time = 1 (s)
-        if time > total_length:
-            time = 1
+            iFrame = 0
+            while True:
+                iFrame += 1
+                ret, frame = vcap.read()
 
-        iFrame = 0
-        while True:
-            iFrame += 1
-            ret, frame = vcap.read()
+                if not ret:
+                    return False
 
-            if not ret:
-                return False
-
-            if round(iFrame/fps_int) > time:
-                cv2.imwrite(output_file, frame)
-                return True
+                if round(iFrame/fps_int) > time:
+                    cv2.imwrite(output_file, frame)
+                    return True
+        else:
+            logging.warn('File not exists -- {}'.format(file_path))
+            return False
     except Exception as e:
-        logging.error('ERROR -- {}'.format(file_path))
-        print(e.__str__)
+        logging.error('ERROR at -- {}'.format(file_path))
+        print(e)
+        return False
 
 
 def auto_generated():
@@ -100,4 +108,4 @@ if __name__ == "__main__":
     fommatExcelFile(os.path.join(base_path, '2_output', 'outputFile.xlsx'))
     logging.info('Write output file -- {}'.format(os.path.join(base_path,
                  '2_output', 'outputFile.xlsx')))
-    os.system("pause")
+    # os.system("pause")
